Ajustes no bot

+++ main.py
@@ -4,7 +4,7 @@ import config
 import consts
 import traceback
 import random
-import youtube_dl
+import youtube_dlc as youtube_dl
 from youtubesearchpython import SearchVideos
 import json
 import asyncio
@@ -13,14 +13,89 @@ import os
 import ctypes
 import ctypes.util
 import datetime
+from twitch import TwitchClient
+from pathlib import Path
+from threading import Thread
+# import nest_asyncio
+
+
+# nest_asyncio.apply()
 
 
 client = discord.Client()
 config.create_table()
 
+twitch_listener_running = False
 queue = {}
 loop_times = -1
 current_volume = 0.2
+
+def extractUrl(song_url):
+    for _ in range(3):
+        try:
+            with youtube_dl.YoutubeDL(consts.YDL_OPTIONS) as ydl:
+                info_dict = ydl.extract_info(song_url, download=False)
+                video_url = info_dict.get("url", None)
+                video_title = info_dict.get('title', None)
+                video_duration = info_dict.get("duration", None)
+                if video_duration is not None:
+                    video_duration = f"{str(video_duration//60).zfill(2)}:{str(video_duration%60).zfill(2)}"
+                
+                return video_title, video_url, video_duration
+        except Exception as error:
+            print(error)
+            return None
+
+async def twitch_listener():
+    twitch_client = TwitchClient(client_id=consts.ENV["TWITCH_CLIENT_ID"])
+
+    while True:
+        existent_users = []
+        if Path("./listened_lives.txt").is_file():
+            _file = open("listened_lives.txt", "r")
+            existent_users = _file.read().split(":")
+            _file.close()
+
+        if existent_users:
+            for user_id in existent_users:
+                stream = twitch_client.streams.get_stream_by_user(user_id)
+                is_live = True if stream is not None else False
+                if is_live:
+                    # await client.get_channel(757390167995711488).send(f"{user_id} is live: {is_live}")
+                    await client.get_channel(757390167995711488).send(
+                        embed=discord.Embed(
+                            color=discord.Color.from_rgb(*dominant_color.to_rgb()),
+                            title=f"Live Advertisement - {stream_name}",
+                            description=stream
+                        )
+                    )
+                    stream_name = stream.get("display_name", None)
+                    stream_url = stream.get("url", None)
+                    stream_profile = stream.get("profile_banner", None)
+                    stream_game = stream.get("game", None)
+
+
+                    if stream_name is None or stream_url is None or stream_profile is None or stream_game is None:
+                        continue
+
+                    _, dominant_color = get_dominant_color.run(stream_url)
+                    dominant_color = discord.Color(value=int(dominant_color, 16))
+                    embed = discord.Embed(
+                        color=discord.Color.from_rgb(*dominant_color.to_rgb()),
+                        title=f"Live Advertisement - {stream_name}",
+                    )
+                    embed.set_image(url=stream_profile)
+                    embed.add_field(
+                        name="Stream URL", value=f"```{stream_url}```", inline=True
+                    )
+                    embed.add_field(
+                        name="Game", value=f"```{stream_game}```", inline=True
+                    )
+
+                    await client.get_channel(757390167995711488).send(
+                        embed=embed
+                    )
+                    print(is_live)
 
 
 def cmd(str):
@@ -33,17 +108,24 @@ def get_main_channel() -> discord.TextChannel:
 
 @client.event
 async def on_ready():
+    global twitch_listener_running
     print(f"We have logged in as {client.user}")
+    if not twitch_listener_running:
+        twitch_listener_running = not twitch_listener_running
+        # asyncio.Task(twitch_listener())
+
+        # loop = asyncio.get_event_loop()
+        # loop.create_task(twitch_listener())
+        # Thread(target=loop.run_forever())
 
 
 @client.event
 async def on_message(message):
     global queue, current_volume, loop_times
-
     if message.author == client.user or message.author.bot:
         return
 
-    _allowed_in_channel_test = ['0188', '6234']
+    _allowed_in_channel_test = ["0188", "6234"]
     if message.channel.name == "channel-test" and (message.author.discriminator not in _allowed_in_channel_test):
         _msg = await message.channel.send(
             embed=discord.Embed(
@@ -71,6 +153,10 @@ async def on_message(message):
         return
     elif (
         not message.content.startswith(cmd(""))
+        and not message.content.startswith("!")
+        and not message.content.startswith("-")
+        and not message.content.startswith("_")
+        and not message.content.startswith("/")
     ) and message.channel.name == "comandos-bot":
         _msg = await message.channel.send(
             embed=discord.Embed(
@@ -190,7 +276,7 @@ async def on_message(message):
             print(traceback.print_exc())
 
             await message.channel.send(f"Error:\n```{error}```")
-    elif message.content.startswith(cmd("loop ")):
+    elif message.content.startswith(cmd("loop")):
         _arg = message.content.split(" ", 1)
         if len(_arg) > 1:
             _arg = message.content.split(" ", 1)[1].strip()
@@ -200,6 +286,11 @@ async def on_message(message):
                 if _arg < 0:
                     raise Exception()
                 loop_times = _arg
+                await message.channel.send(
+                    embed=discord.Embed(
+                        color=0x00FF00, description=f"Loop changed to {loop_times} time(s)"
+                    )
+                )
             except Exception:
                 await message.channel.send(
                     embed=discord.Embed(
@@ -423,7 +514,7 @@ async def on_message(message):
                 message.guild.voice_client.source.volume = current_volume
             else:
                 _voice_client = message.guild.voice_client
-                if _voice_client.is_connected():
+                if _voice_client is not None and _voice_client.is_connected():
                     if queue:
                         queue.clear()
                     coro = _voice_client.disconnect()
@@ -444,31 +535,28 @@ async def on_message(message):
                 )
             )
             yt_url = "https://www.youtube.com/watch?v=mrZC1Jcv0dw"
-            with youtube_dl.YoutubeDL(consts.YDL_OPTIONS) as ydl:
-                song_info = ydl.extract_info(yt_url, download=False)
-                songname = song_info.get("title", None)
-                song_duration = song_info.get("duration", None)
+            song_name, song_url, song_duration = extractUrl(yt_url)
 
-                if not queue.get("queue", None):
-                    queue["queue"] = []
+            if not queue.get("queue", None):
+                queue["queue"] = []
 
-                queue["queue"].insert(0, queue["current"])
-                queue["queue"].insert(
-                    0,
-                    {
-                        "name": songname,
-                        "url": song_info["formats"][0]["url"],
-                        "duration": song_duration,
-                    },
+            queue["queue"].insert(0, queue["current"])
+            queue["queue"].insert(
+                0,
+                {
+                    "name": songname,
+                    "url": song_info["formats"][0]["url"],
+                    "duration": song_duration,
+                },
+            )
+            message.guild.voice_client.stop()
+            await _msg_priorizado.delete()
+            await message.channel.send(
+                embed=discord.Embed(
+                    color=0x00FF00,
+                    description=f"**Playing** `{songname}` - `{song_duration}`",
                 )
-                message.guild.voice_client.stop()
-                await _msg_priorizado.delete()
-                await message.channel.send(
-                    embed=discord.Embed(
-                        color=0x00FF00,
-                        description=f"**Playing** `{songname}` - `{song_duration}`",
-                    )
-                )
+            )
 
             return
         if message.author.voice is None:
@@ -498,6 +586,8 @@ async def on_message(message):
             song_info = ydl.extract_info(yt_url, download=False)
             songname = song_info.get("title", None)
             song_duration = song_info.get("duration", None)
+            if song_duration is not None:
+                song_duration = f"{int(song_duration) // 60}:{int(song_duration) % 60}"
 
             if queue is None or queue.get("current", None) is None:
                 queue = {"current": {}, "queue": []}
@@ -572,7 +662,7 @@ async def on_message(message):
                 message.guild.voice_client.source.volume = current_volume
             else:
                 _voice_client = message.guild.voice_client
-                if _voice_client.is_connected():
+                if _voice_client is not None and _voice_client.is_connected():
                     if queue:
                         queue.clear()
                     coro = _voice_client.disconnect()
@@ -594,25 +684,32 @@ async def on_message(message):
                     color=0x00FF00, description=f"**Searching** `{_arg}`..."
                 )
             )
-            yt = SearchVideos(_arg, offset=1, mode="json", max_results=1)
-            _result = json.loads(yt.result())["search_result"]
+            _result = {}
+            if _arg.startswith('http'):
+                _result['link'] = _arg
+            else:
+                yt = SearchVideos(_arg, offset=1, mode="json", max_results=1)
+                _result = json.loads(yt.result())["search_result"]
+                if not _result:
+                    await searchingMessage.delete()
+                    await message.channel.send(
+                        embed=discord.Embed(
+                            color=0xFF0000,
+                            description=f"Song not found :(",
+                        )
+                    )
+                    return
 
-            if not _result:
-                await searchingMessage.delete()
+                _result = _result[0]
+            song_name, song_url, song_duration = extractUrl(_result["link"])
+            if song_url is None:
                 await message.channel.send(
                     embed=discord.Embed(
                         color=0xFF0000,
-                        description=f"Song not found :(",
+                        description=f"Error on play the song :(",
                     )
-                )
+                ) 
                 return
-            _result = _result[0]
-            song_name = _result["title"]
-            song_duration = _result["duration"]
-            song_url = _result["link"]
-            with youtube_dl.YoutubeDL(consts.YDL_OPTIONS) as ydl:
-                song_url = ydl.extract_info(song_url, download=False)
-                song_url = song_url["formats"][0]["url"]
             if not queue.get("queue", None):
                 queue["queue"] = []
 
@@ -656,24 +753,32 @@ async def on_message(message):
             )
         )
 
-        yt = SearchVideos(_arg, offset=1, mode="json", max_results=1)
-        _result = json.loads(yt.result())["search_result"]
-        if not _result:
-            await searchingMessage.delete()
+        _result = {}
+        if _arg.startswith('http'):
+            _result['link'] = _arg
+        else:
+            yt = SearchVideos(_arg, offset=1, mode="json", max_results=1)
+            _result = json.loads(yt.result())["search_result"]
+            if not _result:
+                await searchingMessage.delete()
+                await message.channel.send(
+                    embed=discord.Embed(
+                        color=0xFF0000,
+                        description=f"Song not found :(",
+                    )
+                )
+                return
+
+            _result = _result[0]
+        song_name, song_url, song_duration = extractUrl(_result["link"])
+        if song_url is None:
             await message.channel.send(
                 embed=discord.Embed(
                     color=0xFF0000,
-                    description=f"Song not found :(",
+                    description=f"Error on play the song :(",
                 )
-            )
+            ) 
             return
-        _result = _result[0]
-        song_name = _result["title"]
-        song_duration = _result["duration"]
-        song_url = _result["link"]
-        with youtube_dl.YoutubeDL(consts.YDL_OPTIONS) as ydl:
-            song_url = ydl.extract_info(song_url, download=False)
-            song_url = song_url["formats"][0]["url"]
 
         if queue is None or queue.get("current", None) is None:
             queue = {"current": {}, "queue": []}
@@ -700,7 +805,7 @@ async def on_message(message):
         )
     elif message.content == cmd("stop"):
         _voice_client = message.guild.voice_client
-        if _voice_client.is_connected():
+        if _voice_client is not None and _voice_client.is_connected():
             if queue:
                 queue.clear()
             await _voice_client.disconnect()
@@ -719,9 +824,25 @@ async def on_message(message):
         await message.channel.send(
             f"<@{user_id}> é {goy_percent_pre}.{goy_percent_pos}% goy! <:56781042_811112372584549_2847201:575829560285986816>"
         )
+    elif message.content == cmd("destiny"):
+        _to_cyberpunk = datetime.datetime(
+            year=2020, month=11, day=10, hour=14, minute=0, second=0
+        )
+
+        _now = datetime.datetime.now()
+
+        _remaining = _to_cyberpunk - _now
+
+        days, seconds = _remaining.days, _remaining.seconds
+        hours = seconds // 3600
+        minutes = (seconds % 3600) // 60
+        seconds = seconds % 60
+
+        msg = f"Tempo até o lançamento da DLC do Destiny {(days)} dias, {hours} horas, {minutes} minutos e {seconds} segundos"
+        await message.channel.send(msg)
     elif message.content == cmd("cyberpunk"):
         _to_cyberpunk = datetime.datetime(
-            year=2020, month=11, day=19, hour=0, minute=1, second=0
+            year=2020, month=12, day=10, hour=0, minute=1, second=0
         )
 
         _now = datetime.datetime.now()
@@ -735,6 +856,37 @@ async def on_message(message):
 
         msg = f"Tempo até o lançamento do cyberpunk {(days)} dias, {hours} horas, {minutes} minutos e {seconds} segundos"
         await message.channel.send(msg)
+    elif message.content.startswith(cmd("live ")):
+        existent_users = []
+        _arg = message.content.split(" ", 1)[1].strip()
+        if Path("./listened_lives.txt").is_file():
+            _file = open("listened_lives.txt", "r")
+            existent_users = _file.read().split(":")
+            _file.close()
+        channel = _arg
+
+        twitch_client = TwitchClient(client_id=consts.ENV["TWITCH_CLIENT_ID"])
+        channels = twitch_client.search.channels(channel, limit=1)
+        if not channels:
+            await message.channel.send(
+                embed=discord.Embed(
+                    color=0xFF0000,
+                    description=f"Channel `{channel}` not found :(",
+                )
+            )
+        else:
+            display_name = channels[0]["display_name"]
+            channel_id = str(channels[0]["id"])
+            await message.channel.send(
+                embed=discord.Embed(
+                    color=0x00FF00,
+                    description=f"Channel `{display_name}` has founded and added to be listened!",
+                )
+            )
+            existent_users.append(channel_id)
+        _file = open("listened_lives.txt", "w")
+        _file.write(":".join(existent_users))
+        _file.close()
     elif message.content == cmd("help"):
         _help = """
         `\level`
